Fix review comments - Add description builder struct
